refactor: Seat 패키지 이동
