+++ app/src/main/java/woowacourse/movie/domain/model/Seat.kt
@@ -1,14 +1,16 @@
 package woowacourse.movie.domain.model
 
+import woowacourse.movie.presentation.model.SeatModel
+
 data class Seat(
     val column: String,
     val row: Int,
     val seatRank: SeatRank,
 )
 
-data class SeatModel(
-    val column: String,
-    val row: Int,
-    val seatRank: SeatRank,
-    var isSelected: Boolean = false,
-)
+fun Seat.toSeatModel(): SeatModel =
+    SeatModel(
+        column = column,
+        row = row,
+        seatRank = seatRank,
+    )
