+++ app/src/main/java/woowacourse/movie/presentation/model/UserSeat.kt
@@ -1,6 +1,5 @@
 package woowacourse.movie.presentation.model
 
-import woowacourse.movie.domain.model.SeatModel
 import java.io.Serializable
 
 data class UserSeat(val seatModels: List<SeatModel>) : Serializable
