+++ app/src/main/java/woowacourse/movie/presentation/ui/seatselection/SeatSelectionContract.kt
@@ -1,10 +1,10 @@
 package woowacourse.movie.presentation.ui.seatselection
 
 import woowacourse.movie.domain.model.Screen
-import woowacourse.movie.domain.model.SeatModel
 import woowacourse.movie.presentation.base.BasePresenter
 import woowacourse.movie.presentation.base.BaseView
 import woowacourse.movie.presentation.model.ReservationInfo
+import woowacourse.movie.presentation.model.SeatModel
 import woowacourse.movie.presentation.model.UserSeat
 
 interface SeatSelectionContract {
